airbrake: load Rake integration only if Rake::Task is defined
Fixes #449 (Airbrake 5 Breaks Debugging)

The problem occurs *only* in RubyMine. For some reason when running
tests with help of it (via the GUI) the `Rake` constant is defined,
but none of the Rake library is loaded. I tried to debug this for a
while, but couldn't find any clues.

The fix is to rely on other constant being present - `Rake::Task`.